split router.rs off from sync.rs
separate the HTTP details from the business logic
